backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
neo4j==5.27.0
openai==1.60.0
tavily-python==0.5.0
gliner2==1.2.3
python-dotenv==1.0.1
httpx==0.28.1
librosa==0.10.2
soundfile==0.13.1
numpy==2.2.3
